# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ParchiFlow - Universal SME Ledger Intelligence & Collection Automation OS
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /reconciliation:
    get:
      operationId: getReconciliationQueue
      tags: [reconciliation]
      summary: Get reconciliation queue
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: issue_type
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Reconciliation items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ReconciliationItem"
...
  /reconciliation/{id}/confirm:
    post:
      operationId: confirmReconciliation
      tags: [reconciliation]
      summary: Confirm a reconciliation match
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReconciliationActionBody"
      responses:
        "200":
          description: Result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /reconciliation/{id}/merge:
    post:
      operationId: mergeReconciliation
      tags: [reconciliation]
      summary: Merge duplicate in reconciliation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReconciliationActionBody"
      responses:
        "200":
          description: Result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /reconciliation/{id}/assign-party:
    post:
      operationId: assignPartyReconciliation
      tags: [reconciliation]
      summary: Assign party to unmatched item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssignPartyBody"
      responses:
        "200":
          description: Result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /reconciliation/{id}/mark-dispute:
    post:
      operationId: markDisputeReconciliation
      tags: [reconciliation]
      summary: Mark item as disputed
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReconciliationActionBody"
      responses:
        "200":
          description: Result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /reconciliation/{id}/ignore:
    post:
      operationId: ignoreReconciliation
      tags: [reconciliation]
      summary: Ignore reconciliation item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReconciliationActionBody"
      responses:
        "200":
          description: Result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /reconciliation/{id}/keep-separate:
    post:
      operationId: keepSeparateReconciliation
      tags: [reconciliation]
      summary: Keep items separate (not duplicates)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReconciliationActionBody"
      responses:
        "200":
          description: Result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # FOLLOW-UPS
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]

    # AUTH SCHEMAS
...
    ReconciliationItem:
      type: object
      properties:
        id:
          type: integer
        businessId:
          type: integer
        sourceType:
          type: string
        issueType:
          type: string
        confidenceScore:
          type: number
        reason:
          type: string
        suggestedAction:
          type: string
        status:
          type: string
        amount:
          type: number
          nullable: true
        partyName:
          type: string
          nullable: true
        suggestedPartyId:
          type: integer
          nullable: true
        referenceNumber:
          type: string
          nullable: true
        utr:
          type: string
          nullable: true
        eventDate:
          type: string
          format: date-time
          nullable: true
        userAction:
          type: string
          nullable: true
        moneyEventId:
          type: integer
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, businessId, sourceType, issueType, confidenceScore, reason, suggestedAction, status, createdAt]
...
    ReconciliationActionBody:
      type: object
      properties:
        note:
          type: string
          nullable: true
      required: []
    AssignPartyBody:
      type: object
      properties:
        partyId:
          type: integer
          nullable: true
        createNewParty:
          type: boolean
        newPartyName:
          type: string
          nullable: true
        note:
          type: string
          nullable: true
      required: []

    # FOLLOW-UP SCHEMAS
...